SensorFusion now uses list of Sensors
- Moved all object detection algorithms into Sensor
- TODO: Make a child class of Sensor, ie lidar or camera to contain
    sensor specific functions
- TODO: fix Measurement and trackmanagement
- Saved vscode window workspace configs
