+++ src/mike_av_stack/scripts/sensor_fusion/configs/sensors.json
@@ -3,6 +3,7 @@
         {
             "type": "sensor.camera.rgb",
             "id": "front",
+            "base_topic": "/carla/ego_vehicle/camera/rgb/front/image_color",
             "x": 2.0, "y": 0.0, "z": 2.0, "roll": 0.0, "pitch": 0.0, "yaw": 0.0,
             "width": 800,
             "height": 600,
@@ -49,6 +50,7 @@
         {
             "type": "sensor.camera.rgb",
             "id": "view",
+            "base_topic": "/carla/ego_vehicle/camera/rgb/view/image_color",
             "x": -4.5, "y": 0.0, "z": 2.8, "roll": 0.0, "pitch": -20.0, "yaw": 0.0,
             "width": 800,
             "height": 600,
@@ -95,6 +97,7 @@
         {
             "type": "sensor.lidar.ray_cast",
             "id": "lidar1",
+            "base_topic": "/carla/ego_vehicle/lidar/lidar1/point_cloud",
             "x": 0.0, "y": 0.0, "z": 2.4, "roll": 0.0, "pitch": 0.0, "yaw": 0.0,
             "range": 50,
             "channels": 64,
@@ -107,6 +110,7 @@
         {
             "type": "sensor.other.radar",
             "id": "front",
+            "base_topic": "/carla/ego_vehicle/radar/front/radar",
             "x": 2.0, "y": 0.0, "z": 1.5, "roll": 0.0, "pitch": 0.0, "yaw": 0.0,
             "horizontal_fov": 30.0,
             "vertical_fov": 10.0,
@@ -116,6 +120,7 @@
         {
             "type": "sensor.other.gnss",
             "id": "gnss1",
+            "base_topic": "/carla/ego_vehicle/gnss/gnss1/fix",
             "x": 1.0, "y": 0.0, "z": 2.0,
             "noise_alt_stddev": 0.0, "noise_lat_stddev": 0.0, "noise_lon_stddev": 0.0,
             "noise_alt_bias": 0.0, "noise_lat_bias": 0.0, "noise_lon_bias": 0.0
@@ -123,6 +128,7 @@
         {
             "type": "sensor.other.imu",
             "id": "imu1",
+            "base_topic": "/carla/ego_vehicle/imu/imu1",
             "x": 2.0, "y": 0.0, "z": 2.0, "roll": 0.0, "pitch": 0.0, "yaw": 0.0,
             "noise_accel_stddev_x": 0.0, "noise_accel_stddev_y": 0.0, "noise_accel_stddev_z": 0.0,
             "noise_gyro_stddev_x": 0.0, "noise_gyro_stddev_y": 0.0, "noise_gyro_stddev_z": 0.0,
