+++ src/mike_av_stack/scripts/sensor_fusion/tracking/measurements.py
@@ -12,6 +12,20 @@
 
 # imports
 import numpy as np
+import rospy
+
+
+from std_msgs.msg import Header
+from sensor_msgs.msg import Image, PointCloud2
+from vision_msgs.msg import BoundingBox3D, ObjectHypothesisWithPose, Detection3D, Detection3DArray
+from geometry_msgs.msg import Pose, Point, Vector3, Quaternion
+import detection.objdet_pcl as pcl
+import detection.objdet_detect as odet
+import numpy as np
+import time
+
+import tools.ros_conversions.transformations as transformations
+import ros_numpy
 
 # add project directory to python path to enable relative imports
 import os
@@ -22,15 +36,34 @@ sys.path.append(os.path.normpath(os.path.join(SCRIPT_DIR, PACKAGE_PARENT)))
 
 class Sensor:
     '''Sensor class including measurement matrix'''
-    def __init__(self, name, calib):
-        self.name = name
-        if name == 'lidar':
+    def __init__(self, configs, configs_track_management = None):
+        self.verbose = True
+        self.configs = configs
+        name = self.configs.type.split('.')[1]
+        self.name = self.configs.type.split('.')[2] if 'other' in name else name
+
+        self.configs.update(odet.load_configs())
+        # Change this so that configs is split up in to each type
+        self.model = odet.create_model(self.configs)
+        self.frame_id = 0
+
+
+        self.pub_detection = rospy.Publisher('/sensor_fusion/detection/' + self.configs.id, Detection3DArray, queue_size=10)
+
+        rospy.Subscriber(self.configs.base_topic, PointCloud2, self.pclCallback)
+
+        if self.name == 'lidar':
             self.dim_meas = 3
             self.sens_to_veh = np.matrix(np.identity((4))) # transformation sensor to vehicle coordinates equals identity matrix because lidar detections are already in vehicle coordinates
             self.fov = [-np.pi/2, np.pi/2] # angle of field of view in radians
         
-        elif name == 'camera':
+        elif self.name == 'camera':
             self.dim_meas = 2
+            if 'calib' not in self.configs :
+                rospy.loginfo('No calibration settings in the sensors.json file. Consider adding them')
+                return
+
+            calib = self.configs.calib
             self.sens_to_veh = np.matrix(calib.extrinsic.transform).reshape(4,4) # transformation sensor to vehicle coordinates
             self.f_i = calib.intrinsic[0] # focal length i-coordinate
             self.f_j = calib.intrinsic[1] # focal length j-coordinate
@@ -97,7 +130,7 @@ class Sensor:
             # END student code
             ############ 
         
-    def get_H(self, x):
+    def get_H(self, x, params):
         # calculate Jacobian H at current x from h(x)
         H = np.matrix(np.zeros((self.dim_meas, params.dim_state)))
         R = self.veh_to_sens[0:3, 0:3] # rotation
@@ -128,27 +161,83 @@ class Sensor:
                                     + R[0,2] * (R[2,0]*x[0] + R[2,1]*x[1] + R[2,2]*x[2] + T[2]) \
                                         / ((R[0,0]*x[0] + R[0,1]*x[1] + R[0,2]*x[2] + T[0])**2))
         return H   
+
+    
+    def imgCallback(self, image):
+        rospy.loginfo('Got an image')
+
+
+    def get_point_cloud_2d(self, pointcloud):
+        # Convert the data from a 1d list of uint8s to a 2d list
+        field_len = len(pointcloud.data)
+
+        point_cloud_2d = np.array([np.array(x.tolist()) for x in ros_numpy.point_cloud2.pointcloud2_to_array(pointcloud)])
+    
+        if self.verbose:
+            print("Shape of pc2d: ", point_cloud_2d.shape, " First element: ", type(point_cloud_2d[0]), point_cloud_2d[0])
+            print("First og: ", pointcloud.data[0], ", ", pointcloud.data[1], ", ", pointcloud.data[2], ", ", pointcloud.data[3])
+            print("height: %d, width: %d, length of data: %d" % (pointcloud.height, pointcloud.width, field_len))
+            for field in pointcloud.fields:
+                print("\tname: ", field.name, "offset: ", field.offset, "datatype: ", field.datatype, "count: ", field.count)
+
+        # TODO: Will need to transform to vehicle coordinate system
+
+        # perform coordinate conversion
+        # xyz_sensor = np.stack([x,y,z,np.ones_like(z)])
+        # xyz_vehicle = np.einsum('ij,jkl->ikl', extrinsic, xyz_sensor)
+        # xyz_vehicle = xyz_vehicle.transpose(1,2,0)
+
+        # transform 3d points into vehicle coordinate system
+        # pcl = xyz_vehicle[ri_range > 0,:3]
+        # pcd = o3d.geometry.PointCloud()
+        # pcd.points = o3d.utility.Vector3dVector(pcl)
+        # o3d.visualization.draw_geometries([pcd])
+
+        return point_cloud_2d
+
+    def pclCallback(self, pointCloud):
+        if self.verbose:
+            rospy.loginfo('Got pointcloud')
+
+        point_cloud_2d = self.get_point_cloud_2d(pointCloud)
+        bev = pcl.bev_from_pcl(point_cloud_2d, self.configs)
+        detections = odet.detect_objects(bev, self.model, self.configs)
+
+        if self.verbose:
+            print(len(detections))
+
+        dets = []
+        for det in detections:
+            d3d = Detection3D()
+            q = transformations.euler_to_quaternion(0, 0, det[7])
+            ori = Quaternion(q[0], q[1], q[2], q[3])
+            d3d.bbox.center.orientation = ori
+            d3d.bbox.center.position.x = det[1]
+            d3d.bbox.center.position.y = det[2]
+            d3d.bbox.center.position.z = det[3]
+            d3d.bbox.size.x = det[4]
+            d3d.bbox.size.y = det[5]
+            d3d.bbox.size.z = det[6]
+            dets.append(d3d)
+
+        time_now = time.time_ns()
+        header = Header()
+        self.frame_id += 1
+        header.frame_id = self.frame_id 
+        header.stamp.secs = int(time_now / 10e9)
+        header.stamp.nsecs = time_now - (header.stamp.secs * 10e9)
+        detection3DArray = Detection3DArray()
+        detection3DArray.header = Header() 
+        detection3DArray.header
+        detection3DArray.detections = dets
+        self.pub_detection.publish(detection3DArray)
+    
         
-    def generate_measurement(self, num_frame, z, meas_list):
-        # generate new measurement from this sensor and add to measurement list
-        ############
-        # Step 4: remove restriction to lidar in order to include camera as well
-        ############
-        
-        meas = Measurement(num_frame, z, self)
-        meas_list.append(meas)
-        return meas_list
-        
-        ############
-        # END student code
-        ############ 
-        
-        
-################### 
+
         
 class Measurement:
     '''Measurement class including measurement values, covariance, timestamp, sensor'''
-    def __init__(self, num_frame, z, sensor):
+    def __init__(self, num_frame, z, sensor, params):
         # create measurement object
         self.t = (num_frame - 1) * params.dt # time
         self.sensor = sensor # sensor that generated this measurement
